+++ scripts/test_nearlib.sh
@@ -9,7 +9,7 @@ trap 'pkill -15 -P $NEAR_PID' 0
 
 # Run nearlib tests
 rm -rf nearlib
-git clone --single-branch --branch master https://github.com/nearprotocol/nearlib.git nearlib
+git clone --single-branch --branch nep8 https://github.com/nearprotocol/nearlib.git nearlib
 cd nearlib
 export NEAR_PROTOS_DIR="../core/protos/protos"
 export HELLO_WASM_PATH="../tests/hello.wasm"
