+++ test-utils/testlib/src/user/mod.rs
@@ -7,13 +7,14 @@ use near_primitives::crypto::signature::PublicKey;
 use near_primitives::crypto::signer::EDSigner;
 use near_primitives::hash::CryptoHash;
 use near_primitives::receipt::{Receipt, ReceiptInfo};
+use near_primitives::rpc::FinalTransactionResult;
 use near_primitives::rpc::{
     AccessKeyView, AccountView, BlockView, CryptoHashView, ViewStateResult,
 };
 use near_primitives::transaction::{
     Action, AddKeyAction, CreateAccountAction, DeleteAccountAction, DeleteKeyAction,
-    DeployContractAction, FinalTransactionResult, FunctionCallAction, SignedTransaction,
-    StakeAction, TransactionResult, TransferAction,
+    DeployContractAction, FunctionCallAction, SignedTransaction, StakeAction, TransactionResult,
+    TransferAction,
 };
 use near_primitives::types::{AccountId, Balance, Gas, MerkleHash};
